Misc sccz80 fixes and opts (#698)
* Promotion issues with casts.

Optimisations.

Breaks Issue_678_compound.c and paramsoffset tests

* Remove the duplicate int2long_s (paramsoffset) and fix results for
Issue_678_compound - the new result is correct.

* Add a quick case for multiplying by 8192 as well

* Some more rules

* Use the new disassembler in the debugger.

Add +zxn MMU paging (in a very casual way)

* We need sys/compiler at the start or it won't work if included
first of all

* Set the file mode when we open the file otherwise perms go awol

* Inline int2long_s as well

* Add a hybrid mode

* Don't use C99

* Fix a rule that was breaking:

ld	hl,(_somevariable)
ld	(hl),0
inc	hl
ld	(hl),0

and replacing with

ld	hl,0
ld	((_somevariable)),hl

* Correct the rule correction

* ZXnext mmu debugging, correct register initialise.

* windows compatibility

* Issue #653: Fix empty functions not generating a ret if the previous function
ended in return.

Some more opt rule

Remove debug from ticks

* Add arch/test.h that has the zxnext mmu operations - the zxnext
memory is partially support by ticks

* Restore +x so that it works on linux

* Implement nextreg in ticks

* Switch to z80-zxn (without compiler support)

* Remove erroneous comment

+++ include/intrinsic.h
@@ -4,6 +4,8 @@
 #ifndef __INTRINSIC_H__
 #define __INTRINSIC_H__
 
+#include <sys/compiler.h>
+
 #ifdef __SDCC
 
 // SDCC Compiles
@@ -94,7 +96,7 @@ extern void __LIB__ intrinsic_exx(void) __smallc;
 extern void __LIB__ *intrinsic_return_bc(void) __smallc;
 extern void __LIB__ *intrinsic_return_de(void) __smallc;
 
-extern unsigned int __LIB__ intrinsic_swap_endian_16(unsigned long n) __smallc __z88dk_fastcall;
+extern unsigned int __LIB__ intrinsic_swap_endian_16(unsigned int n) __smallc __z88dk_fastcall;
 extern unsigned long __LIB__ intrinsic_swap_endian_32(unsigned long n) __smallc __z88dk_fastcall;
 extern unsigned long __LIB__ intrinsic_swap_word_32(unsigned long n) __smallc __z88dk_fastcall;
